Align FileSystemAbstraction with java FileSystem
After the move to NIO there were a few differences left:
 * `mkdir` would return a boolean to indicate success
 * `getFileSize` would never fail, and return 0 on failure
 * `deleteFile` would return a boolean to indicate success
 * `listFiles` would never fail, returning `null` on both empty and failure
 * `lastModifiedTime` would never fail, and return 0 on failure.

 All of the above will now throw the appropriate exception when an error occurs.

+++ community/batch-insert/src/main/java/org/neo4j/batchinsert/internal/BatchInserterImpl.java
@@ -629,7 +629,10 @@ public class BatchInserterImpl implements BatchInserter
     private void rebuildCounts( PageCacheTracer cacheTracer, MemoryTracker memoryTracker ) throws IOException
     {
         Path countsStoreFile = databaseLayout.countStore();
-        fileSystem.deleteFile( countsStoreFile );
+        if ( fileSystem.fileExists( countsStoreFile ) )
+        {
+            fileSystem.deleteFile( countsStoreFile );
+        }
         CountsComputer initialCountsBuilder =
                 new CountsComputer( neoStores, pageCache, cacheTracer, databaseLayout, memoryTracker, logService.getInternalLog( getClass() ) );
         try ( GBPTreeCountsStore countsStore = new GBPTreeCountsStore( pageCache, databaseLayout.countStore(), fileSystem, immediate(),
